# backend/docs/swagger.yaml
definitions:
  auth.LoginUserRequest:
    properties:
      masterKeyHash:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - masterKeyHash
    - password
    - username
    type: object
  auth.LoginUserResponse:
    properties:
      expiresAt:
        type: integer
      name:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  auth.RecoverRequest:
    properties:
      password:
        type: string
      recoveryKey:
        type: string
      username:
        type: string
    required:
    - password
    - recoveryKey
    - username
    type: object
  auth.RecoverResponse:
    properties:
      encryptedMasterKey:
        type: string
    type: object
  auth.RegisterUserRequest:
    properties:
      masterKeyHash:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - masterKeyHash
    - name
    - password
    - username
    type: object
  auth.RegisterUserResponse:
    properties:
      expiresAt:
        type: integer
      name:
        type: string
      recoveryKey:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  auth.SetupRecoveryRequest:
    properties:
      encryptedMasterKey:
        type: string
    required:
    - encryptedMasterKey
    type: object
...
  models.Error:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      statusCode:
        type: integer
    type: object
  models.GenericResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: API server for FlySsh
  title: FlySsh API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with username, password and masterKeyHash.
        The client derives authKey = PBKDF2(masterKey, email) and sends it as masterKeyHash.
        The server verifies it against the stored bcrypt hash.
      parameters:
      - description: Login details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Login a user
      tags:
      - auth
  /auth/recover:
    post:
      consumes:
      - application/json
      description: Verify username, password, and recovery key, then return the encrypted
        master key blob. The client decrypts it locally. No JWT is issued — call /auth/login
        afterward.
      parameters:
      - description: Recovery credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RecoverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.RecoverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
      summary: Recover encrypted master key
      tags:
      - auth
  /auth/recovery/setup:
    post:
      consumes:
      - application/json
      description: After registration, the client encrypts the raw master key using
        the recovery key (AES-GCM, client-side) and sends the encrypted blob to the
        server for storage.
      parameters:
      - description: Encrypted master key blob
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SetupRecoveryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Store encrypted master key for recovery
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account. The client must generate a master key
        locally, derive an authKey via PBKDF2(masterKey, email), and send the authKey
        as masterKeyHash. The server never sees the raw master key.
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"